로그인한 사용자와 게시글을 삭제할 사용자를 검사하여 같을 때 게시글 삭제

+++ src/com/javaex/controller/BoardServlet.java
@@ -61,10 +61,13 @@ public class BoardServlet extends HttpServlet {
 			
 		} else if ("delete".equals(actionName)) {
 			int boardNo = Integer.parseInt(request.getParameter("boardNo"));
-	
-			BoardDao dao = new BoardDao();
-			dao.delete(boardNo);
+			String authUser = request.getParameter("auth");
+			String user = request.getParameter("user");
 			
+			if(user.equals(authUser)) {
+				BoardDao dao = new BoardDao();
+				dao.delete(boardNo);
+			}
 			response.sendRedirect("/mysite/bs");
 
 		} else if ("write".equals(actionName)) {
